Rename Hyrule Field Speed Hack to Hyrule Field Minimap Speedhack
We need "minimap" in the name to describe what the hack actually affects.
It only says "Hyrule Field" because it happened to improve Hyrule Field
performance more than other placces, not because it's specific to it.
Still, I'm keeping "Hyrule Field" in the name so people who are following
instructions that say "Hyrule Field Speed Hack" won't get too confused.