SI: Fix bugs with input for movies
GetPadStatus() causes input to be read from the movie, but gets called
twice in the RunBuffer path for CMD_DIRECT- once at the start, and once
within GetData(). This isn't a problem when polled input wasn't going
through the "normal" command mechanism, but now the result from the first
GetPadStatus is kept and reused.