Android: Set orientation to auto by default
When I added the orientation lock setting and set it to landscape
by default, it was because I wanted to prevent the screen from
unintentionally rotating when playing games where you have to tilt
the Wii Remote heavily, like the cow minigame in Wii Play. However,
playing those kinds of games is kind of a bad experience anyway
because all the tilting makes it hard to see the screen, and they
are also far from the majority of games. I think it's best to allow
rotation by default, providing the smoothest experience to people
who just want to play some GameCube games and who perhaps want to
use portrait mode or have a gamepad that expects reverse landscape.