JitArm64: Add analysis for ppcState LDP/STP
Making better use of LDP/STP when accessing ppcState will both reduce
how many cycles we have to spend on ppcState access in each block and
reduce code size. It does increase register pressure, but with how many
registers we have on AArch64, that shouldn't be much of a problem.

This commit adds the new bitsets load_pairs and store_pairs to
BlockRegStats, which indicate which registers should be treated as a
pair when loading and storing to ppcState. The commits after this one
will add code that reads these bitsets to determine when to use LDP/STP.

x64 doesn't have instructions that can load/store a pair of registers,
so the new PPCAnalyst code is ifdef'd for AArch64.