JitArm64: Remove the "a bit annoying" comment
This comment doesn't make sense in the context of JitArm64.

In Jit64, if the implementation of a guest instruction simply tries to
use a guest register, the register cache will by default emit an
instruction that accesses the ppcState struct directly instead of
loading the value into a host register. This is good to do for guest
registers that will only be used once, but for guest registers that will
be used twice it's a bit inefficient, so before each instruction we run
PreloadRegisters to load those registers which will be used by the
current instruction and will be used at least one more time later.

None of that applies to JitArm64. When the implementation of a guest
instruction tries to access a guest register, the register cache always
loads it into a host register, because the only instructions AArch64 has
that access memory are the load/store instructions. Therefore there is
no PreloadRegisters in JitArm64, and the "a bit annoying" path is the
normal path.

With that said, maybe the comment is referring to how registers can have
gotten loaded when emitting code for an earlier guest instruction... But
I don't think that's much of a reason to keep the comment around. It
doesn't add much useful information.