Jit: Handle CR SO accurately in blocks that set overflow
Normally, we don't care about setting the SO bit of condition registers
accurately in the JITs, because it takes time and no game seems to rely
on it. But if we notice that a block sets the overflow flags in XER, we
might as well take the time to emulate the CR SO bit accurately.