Jit: Fix "skip redundant flushes" for skipped instructions
Normally when an instruction is skipped (for instance due to branch
merging or BLR optimization), the registers that would have been flushed
at the end of that instruction will instead be flushed at the end of the
next instruction, which is maybe not perfect, but usually good enough.
However, since the addition of the "skip redundant flushes" logic in
fd511a689f, Dolphin has accidentally skipped flushing those registers,
which creates unnecessary register pressure. This commit restores the
old behavior.